requirements.txt
kafka
paho-mqtt
psycopg2
boto3
